Crate rustc_save_analysis

Unstable (rustc_private)

Modules

external_data 	 [Unstable]

recorder 	 [Unstable]

span_utils 	 [Unstable]

Structs

CrateData 	 [Unstable]

CratePreludeData 	 [Unstable]
Data for the prelude of a crate.

CsvDumper 	 [Unstable]

DumpVisitor 	 [Unstable]

EnumData 	 [Unstable]
Data for enum declarations.

ExternCrateData 	 [Unstable]
Data for extern crates.

ExternalCrateData 	 [Unstable]
Data for external crates in the prelude of a crate.

FunctionCallData 	 [Unstable]
Data about a function call.

FunctionData 	 [Unstable]
Data for all kinds of functions and methods.

FunctionRefData 	 [Unstable]
Data about a function call.

ImplData 	 [Unstable]

ImplData2 	 [Unstable]

InheritanceData 	 [Unstable]

JsonDumper 	 [Unstable]

MacroData 	 [Unstable]
Data about a macro declaration.

MacroUseData 	 [Unstable]
Data about a macro use.

MethodCallData 	 [Unstable]
Data about a method call.

MethodData 	 [Unstable]
Data for method declarations (methods with a body are treated as functions).

ModData 	 [Unstable]
Data for modules.

ModRefData 	 [Unstable]
Data for a reference to a module.

SaveContext 	 [Unstable]

StructData 	 [Unstable]

StructVariantData 	 [Unstable]

TraitData 	 [Unstable]

TupleVariantData 	 [Unstable]

TypeDefData 	 [Unstable]
Data for a typedef.

TypeRefData 	 [Unstable]
Data for a reference to a type or trait.

UseData 	 [Unstable]

UseGlobData 	 [Unstable]

VariableData 	 [Unstable]
Data for local and global variables (consts and statics).

VariableRefData 	 [Unstable]
Data for the use of some item (e.g., the use of a local variable, which will refer to that variables declaration (by ref_id)).

Enums

Data 	 [Unstable]
Data for any entity in the Rust language. The actual data contained varies with the kind of entity being queried. See the nested structs for details.

Format 	 [Unstable]

Traits

Dump 	 [Unstable]

Functions

generated_code 	 [Unstable]

process_crate 	 [Unstable]