Module rustc_save_analysis::external_data

🔬 This is a nightly-only experimental API. (rustc_private)

Structs

Attribute 	 [
Experimental
]
Represent an arbitrary attribute on a code element

CratePreludeData 	 [
Experimental
]

EnumData 	 [
Experimental
]
Data for enum declarations.

ExternCrateData 	 [
Experimental
]
Data for extern crates.

FunctionCallData 	 [
Experimental
]
Data about a function call.

FunctionData 	 [
Experimental
]
Data for all kinds of functions and methods.

FunctionRefData 	 [
Experimental
]
Data about a function call.

ImplData 	 [
Experimental
]

InheritanceData 	 [
Experimental
]

MacroData 	 [
Experimental
]
Data about a macro declaration.

MacroUseData 	 [
Experimental
]
Data about a macro use.

MethodCallData 	 [
Experimental
]
Data about a method call.

MethodData 	 [
Experimental
]
Data for method declarations (methods with a body are treated as functions).

ModData 	 [
Experimental
]
Data for modules.

ModRefData 	 [
Experimental
]
Data for a reference to a module.

Signature 	 [
Experimental
]

SpanData 	 [
Experimental
]

StructData 	 [
Experimental
]

StructVariantData 	 [
Experimental
]

TraitData 	 [
Experimental
]

TupleVariantData 	 [
Experimental
]

TypeDefData 	 [
Experimental
]
Data for a typedef.

TypeRefData 	 [
Experimental
]
Data for a reference to a type or trait.

UseData 	 [
Experimental
]

UseGlobData 	 [
Experimental
]

VariableData 	 [
Experimental
]
Data for local and global variables (consts and statics).

VariableRefData 	 [
Experimental
]
Data for the use of some item (e.g., the use of a local variable, which will refer to that variables declaration (by ref_id)).

Traits

Lower 	 [
Experimental
]

Functions

make_def_id 	 [
Experimental
]

null_def_id 	 [
Experimental
]