Struct rustc_save_analysis::span_utils::SpanUtils

pub struct SpanUtils<'a> {
    pub sess: &'a Session,
    pub err_count: Cell<isize>,
}

Unstable (rustc_private)

Fields
sess: &'a Session
Unstable (rustc_private)
err_count: Cell<isize>
Unstable (rustc_private)

Methods

impl<'a> SpanUtils<'a>

fn new(sess: &'a Session) -> SpanUtils<'a>

fn make_path_string(file_name: &str) -> String

fn make_sub_span(&self, span: Span, sub_span: Option<Span>) -> Option<Span>

fn snippet(&self, span: Span) -> String

fn retokenise_span(&self, span: Span) -> StringReader<'a>

fn span_for_last_ident(&self, span: Span) -> Option<Span>

fn span_for_first_ident(&self, span: Span) -> Option<Span>

fn sub_span_for_meth_name(&self, span: Span) -> Option<Span>

fn sub_span_for_type_name(&self, span: Span) -> Option<Span>

fn spans_with_brackets(&self, span: Span, nesting: isize, limit: isize) -> Vec<Span>

fn sub_span_before_token(&self, span: Span, tok: Token) -> Option<Span>

fn sub_span_of_token(&self, span: Span, tok: Token) -> Option<Span>

fn sub_span_after_keyword(&self, span: Span, keyword: Keyword) -> Option<Span>

fn sub_span_after_token(&self, span: Span, tok: Token) -> Option<Span>

fn spans_for_path_segments(&self, path: &Path) -> Vec<Span>

fn spans_for_ty_params(&self, span: Span, number: isize) -> Vec<Span>

fn report_span_err(&self, kind: &str, span: Span)

fn span_for_macro_def_name(&self, span: Span) -> Option<Span>

fn span_for_macro_use_name(&self, span: Span) -> Option<Span>

fn filter_generated(&self, sub_span: Option<Span>, parent: Span) -> bool

Unstable (rustc_private)

Return true if the span is generated code, and it is not a subspan of the root callsite.

Used to filter out spans of minimal value, such as references to macro internal variables.