Struct rustc_save_analysis::CallbackHandler

pub struct CallbackHandler<'b> {
    pub callback: &'b mut FnMut(&Analysis),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Call a callback with the results of save-analysis.

Fields
callback: &'b mut FnMut(&Analysis)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?