Struct rustc_save_analysis::DumpHandler

pub struct DumpHandler<'a> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Dump the save-analysis results to a file.

Methods

impl<'a> DumpHandler<'a>

pub fn new(odir: Option<&'a Path>, cratename: &str) -> DumpHandler<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?