Struct rustc_save_analysis::dump_visitor::DumpVisitor

pub struct DumpVisitor<'tcx> {
    pub save_ctxt: SaveContext<'tcx>,
    tcx: TyCtxt<'tcx>,
    dumper: Dumper,
    span: SpanUtils<'tcx>,
}

Fields
save_ctxt: SaveContext<'tcx>tcx: TyCtxt<'tcx>dumper: Dumperspan: SpanUtils<'tcx>
Implementations

impl<'tcx> DumpVisitor<'tcx>

pub fn new(save_ctxt: SaveContext<'tcx>) -> DumpVisitor<'tcx>

pub fn analysis(&self) -> &Analysis

fn nest_typeck_results<F>(&mut self, item_def_id: LocalDefId, f: F)where
F: FnOnce(&mut Self),

fn span_from_span(&self, span: Span) -> SpanData

fn lookup_def_id(&self, ref_id: HirId) -> Option<DefId>

pub fn dump_crate_info(&mut self, name: &str)

pub fn dump_compilation_options(&mut self, input: &Input, crate_name: &str)

fn write_segments(
&mut self,
segments: impl IntoIterator<Item = &'tcx PathSegment<'tcx>>
)

fn write_sub_paths(&mut self, path: &'tcx Path<'tcx>)

fn write_sub_paths_truncated(&mut self, path: &'tcx Path<'tcx>)

fn process_formals(&mut self, formals: &'tcx [Param<'tcx>], qualname: &str)

fn process_method(
&mut self,
sig: &'tcx FnSig<'tcx>,
body: Option<BodyId>,
def_id: LocalDefId,
ident: Ident,
generics: &'tcx Generics<'tcx>,
span: Span
)

fn process_struct_field_def(
&mut self,
field: &'tcx FieldDef<'tcx>,
parent_id: HirId
)

fn process_generic_params(
&mut self,
generics: &'tcx Generics<'tcx>,
prefix: &str,
id: HirId
)

fn process_fn(
&mut self,
item: &'tcx Item<'tcx>,
decl: &'tcx FnDecl<'tcx>,
_header: &'tcx FnHeader,
ty_params: &'tcx Generics<'tcx>,
body: BodyId
)

fn process_static_or_const_item(
&mut self,
item: &'tcx Item<'tcx>,
typ: &'tcx Ty<'tcx>,
expr: &'tcx Expr<'tcx>
)

fn process_assoc_const(
&mut self,
def_id: LocalDefId,
ident: Ident,
typ: &'tcx Ty<'tcx>,
expr: Option<&'tcx Expr<'tcx>>,
parent_id: DefId,
attrs: &'tcx [Attribute]
)

fn process_struct(
&mut self,
item: &'tcx Item<'tcx>,
def: &'tcx VariantData<'tcx>,
ty_params: &'tcx Generics<'tcx>
)

fn process_enum(
&mut self,
item: &'tcx Item<'tcx>,
enum_definition: &'tcx EnumDef<'tcx>,
ty_params: &'tcx Generics<'tcx>
)

fn process_impl(&mut self, item: &'tcx Item<'tcx>, impl_: &'tcx Impl<'tcx>)

fn process_trait(
&mut self,
item: &'tcx Item<'tcx>,
generics: &'tcx Generics<'tcx>,
trait_refs: GenericBounds<'tcx>,
methods: &'tcx [TraitItemRef]
)

fn process_mod(&mut self, item: &'tcx Item<'tcx>)

fn dump_path_ref(&mut self, id: HirId, path: &QPath<'tcx>)

fn dump_path_segment_ref(&mut self, id: HirId, segment: &PathSegment<'tcx>)

fn process_path(&mut self, id: HirId, path: &QPath<'tcx>)

fn process_struct_lit(
&mut self,
ex: &'tcx Expr<'tcx>,
path: &'tcx QPath<'tcx>,
fields: &'tcx [ExprField<'tcx>],
variant: &'tcx VariantDef,
rest: Option<&'tcx Expr<'tcx>>
)

fn process_method_call(
&mut self,
ex: &'tcx Expr<'tcx>,
seg: &'tcx PathSegment<'tcx>,
receiver: &'tcx Expr<'tcx>,
args: &'tcx [Expr<'tcx>]
)

fn process_pat(&mut self, p: &'tcx Pat<'tcx>)

fn process_var_decl(&mut self, pat: &'tcx Pat<'tcx>)

fn process_macro_use(&mut self, _span: Span)

Extracts macro use and definition information from the AST node defined by the given NodeId, using the expansion information from the node’s span.

If the span is not macro-generated, do nothing, else use callee and callsite spans to record macro definition and use data, using the mac_uses and mac_defs sets to prevent multiples.

fn process_trait_item(
&mut self,
trait_item: &'tcx TraitItem<'tcx>,
trait_id: DefId
)

fn process_impl_item(&mut self, impl_item: &'tcx ImplItem<'tcx>, impl_id: DefId)

pub(crate) fn process_crate(&mut self)

fn process_bounds(&mut self, bounds: GenericBounds<'tcx>)