Enum rustc_save_analysis::Data

pub enum Data {
    EnumData(EnumData),
    ExternCrateData(ExternCrateData),
    FunctionCallData(FunctionCallData),
    FunctionData(FunctionData),
    FunctionRefData(FunctionRefData),
    ImplData(ImplData2),
    InheritanceData(InheritanceData),
    MacroData(MacroData),
    MacroUseData(MacroUseData),
    MethodCallData(MethodCallData),
    MethodData(MethodData),
    ModData(ModData),
    ModRefData(ModRefData),
    StructData(StructData),
    StructVariantDat(StructVariantData),
    TraitData(TraitData),
    TupleVariantData(TupleVariantData),
    TypeDefData(TypeDefData),
    TypeRefData(TypeRefData),
    UseData(UseData),
    UseGlobData(UseGlobData),
    VariableData(VariableData),
    VariableRefData(VariableRefData),
}

Unstable (rustc_private)

Data for any entity in the Rust language. The actual data contained varies with the kind of entity being queried. See the nested structs for details.

Variants
EnumData(EnumData)
Unstable (rustc_private)

Data for Enums.

ExternCrateData(ExternCrateData)
Unstable (rustc_private)

Data for extern crates.

FunctionCallData(FunctionCallData)
Unstable (rustc_private)

Data about a function call.

FunctionData(FunctionData)
Unstable (rustc_private)

Data for all kinds of functions and methods.

FunctionRefData(FunctionRefData)
Unstable (rustc_private)

Data about a function ref.

ImplData(ImplData2)
Unstable (rustc_private)

Data for impls.

InheritanceData(InheritanceData)
Unstable (rustc_private)

Data for trait inheritance.

MacroData(MacroData)
Unstable (rustc_private)

Data about a macro declaration.

MacroUseData(MacroUseData)
Unstable (rustc_private)

Data about a macro use.

MethodCallData(MethodCallData)
Unstable (rustc_private)

Data about a method call.

MethodData(MethodData)
Unstable (rustc_private)

Data for method declarations (methods with a body are treated as functions).

ModData(ModData)
Unstable (rustc_private)

Data for modules.

ModRefData(ModRefData)
Unstable (rustc_private)

Data for a reference to a module.

StructData(StructData)
Unstable (rustc_private)

Data for a struct declaration.

StructVariantDat(StructVariantData)
Unstable (rustc_private)

Data for a struct variant.

TraitData(TraitData)
Unstable (rustc_private)

Data for a trait declaration.

TupleVariantData(TupleVariantData)
Unstable (rustc_private)

Data for a tuple variant.

TypeDefData(TypeDefData)
Unstable (rustc_private)

Data for a typedef.

TypeRefData(TypeRefData)
Unstable (rustc_private)

Data for a reference to a type or trait.

UseData(UseData)
Unstable (rustc_private)

Data for a use statement.

UseGlobData(UseGlobData)
Unstable (rustc_private)

Data for a global use statement.

VariableData(VariableData)
Unstable (rustc_private)

Data for local and global variables (consts and statics), and fields.

VariableRefData(VariableRefData)
Unstable (rustc_private)

Data for the use of some variable (e.g., the use of a local variable, which will refere to that variables declaration).