Module rustc_save_analysis::external_data

Unstable (rustc_private)

Structs

CratePreludeData 	 [Unstable]

EnumData 	 [Unstable]
Data for enum declarations.

ExternCrateData 	 [Unstable]
Data for extern crates.

FunctionCallData 	 [Unstable]
Data about a function call.

FunctionData 	 [Unstable]
Data for all kinds of functions and methods.

FunctionRefData 	 [Unstable]
Data about a function call.

ImplData 	 [Unstable]

InheritanceData 	 [Unstable]

MacroData 	 [Unstable]
Data about a macro declaration.

MacroUseData 	 [Unstable]
Data about a macro use.

MethodCallData 	 [Unstable]
Data about a method call.

MethodData 	 [Unstable]
Data for method declarations (methods with a body are treated as functions).

ModData 	 [Unstable]
Data for modules.

ModRefData 	 [Unstable]
Data for a reference to a module.

SpanData 	 [Unstable]

StructData 	 [Unstable]

StructVariantData 	 [Unstable]

TraitData 	 [Unstable]

TupleVariantData 	 [Unstable]

TypeDefData 	 [Unstable]
Data for a typedef.

TypeRefData 	 [Unstable]
Data for a reference to a type or trait.

UseData 	 [Unstable]

UseGlobData 	 [Unstable]

VariableData 	 [Unstable]
Data for local and global variables (consts and statics).

VariableRefData 	 [Unstable]
Data for the use of some item (e.g., the use of a local variable, which will refer to that variables declaration (by ref_id)).

Traits

Lower 	 [Unstable]

Functions

null_def_id 	 [Unstable]