Struct rustc_save_analysis::SaveContext

pub struct SaveContext<'l, 'tcx: 'l> {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl<'l, 'tcx: 'l> SaveContext<'l, 'tcx>

fn new(tcx: TyCtxt<'l, 'tcx, 'tcx>) -> SaveContext<'l, 'tcx>

fn from_span_utils(tcx: TyCtxt<'l, 'tcx, 'tcx>, span_utils: SpanUtils<'tcx>) -> SaveContext<'l, 'tcx>

fn get_external_crates(&self) -> Vec<CrateData>

fn get_item_data(&self, item: &Item) -> Option<Data>

fn get_field_data(&self, field: &StructField, scope: NodeId) -> Option<VariableData>

fn get_method_data(&self, id: NodeId, name: Name, span: Span) -> Option<FunctionData>

fn get_trait_ref_data(&self, trait_ref: &TraitRef, parent: NodeId) -> Option<TypeRefData>

fn get_expr_data(&self, expr: &Expr) -> Option<Data>

fn get_path_data(&self, id: NodeId, path: &Path) -> Option<Data>

fn get_field_ref_data(&self, field_ref: &Field, variant: VariantDef, parent: NodeId) -> Option<VariableRefData>

fn get_macro_use_data(&self, span: Span, id: NodeId) -> Option<MacroUseData>

Unstable (rustc_private)

Attempt to return MacroUseData for any AST node.

For a given piece of AST defined by the supplied Span and NodeId, returns None if the node is not macro-generated or the span is malformed, else uses the expansion callsite and callee to return some MacroUseData.

fn get_data_for_id(&self, _id: &NodeId) -> Data

fn enclosing_scope(&self, id: NodeId) -> NodeId