Struct rustc_save_analysis::SaveContext

pub struct SaveContext<'l, 'tcx: 'l> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'l, 'tcx: 'l> SaveContext<'l, 'tcx>

pub fn get_external_crates(&self) -> Vec<ExternalCrateData>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_extern_item_data(&self, item: &ForeignItem) -> Option<Data>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_item_data(&self, item: &Item) -> Option<Data>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_field_data(&self, field: &StructField, scope: NodeId) -> Option<Def>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_method_data(&self, id: NodeId, name: Name, span: Span) -> Option<Def>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_trait_ref_data(&self, trait_ref: &TraitRef) -> Option<Ref>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_expr_data(&self, expr: &Expr) -> Option<Data>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_path_def(&self, id: NodeId) -> HirDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_path_data(&self, id: NodeId, path: &Path) -> Option<Ref>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_field_ref_data(
&self,
field_ref: &Field,
variant: &VariantDef
) -> Option<Ref>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn get_macro_use_data(&self, span: Span) -> Option<MacroRef>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Attempt to return MacroRef for any AST node.

For a given piece of AST defined by the supplied Span and NodeId, returns None if the node is not macro-generated or the span is malformed, else uses the expansion callsite and callee to return some MacroRef.